User login system and validation created
We can register a new user with /signup route.

We can login with /login route, then get back a json web token to identify us the next times.

When registering a new user, we check that:
- The password has a strong OWASP level
- The password is not compromised (haveibeenpwned)
- The password has between 8 and 64 characters.

Then on user creation or modification, if the password is created or modified, we hash it with the argon2 algorithm.