Add GDPR routes:
The user can export its data.
The user can update his password.
The user can report a sauce or a user.
The user can delete his account.

Better ObjectId verification on routes like something/:id, now if the id is not a valid ObjectId, we skip to the next middleware.

Refactor many methods to async / await for a better readability.

Add a new report field to Users and Sauces schemas.

Debug the handleLike method because it was not working as expected.